test: add validation tests for creating new users

+++ part4/bloglist/bloglist_backend/tests/users_api.test.js
@@ -8,6 +8,10 @@ const bcrypt = require("bcrypt");
 
 const api = supertest(app);
 
+if (process.env.NODE_ENV !== "test") {
+  throw new Error("Tests must run on the test database!");
+}
+
 beforeEach(async () => {
   await User.deleteMany({});
